Fix some warnings when using eigen_vector for the test vector.
configure.ac: Comments for auto-tool commands that will result in many makefile.in changes.
makefile.in: automatic change from corresponding makefile.am.
test_one.sh.in: use cppad_cxx_flags to determine if c++11.
makefile.in: automatic change from corresponding makefile.am.
exp_forward.omh: Fix z^{(0)} to be correct for both cases.
makefile.in: automatic change from corresponding makefile.am.

+++ cppad/local/color_general.hpp
@@ -114,8 +114,8 @@ void color_general_cppad(
 	size_t m = pattern.n_set();
 	size_t n = pattern.end();
 
-	CPPAD_ASSERT_UNKNOWN( col.size()   == K );
-	CPPAD_ASSERT_UNKNOWN( color.size() == m );
+	CPPAD_ASSERT_UNKNOWN( size_t( col.size() )   == K );
+	CPPAD_ASSERT_UNKNOWN( size_t( color.size() ) == m );
 
 	// We define the set of rows, columns, and pairs that appear
 	// by the set ( row[k], col[k] ) for k = 0, ... , K-1.
